adding the timer
